split code to multiple headers
