--- components/mds/mds_cpp/c
@@ -1,3 +1,0 @@
-#! /bin/bash
-g++ $1.cpp -o $1 -std=c++11 -O2 -DARMA_DONT_USE_WRAPPER -DARMA_USE_BLAS -DARMA_USE_LAPACK -lopenblas -llapack
-
