+++ components/mds/mds_cpp/qhull_2d.cpp
@@ -41,21 +41,3 @@ mat qhull2D(mat sample) {
     } else
         return sample;
 }
-
-// test
-// int main() {
-//     mat a = {
-//         {0.46139225,  0.01608542},
-//         {0.93274524,  0.26489567},
-//         {0.32466116,  0.86086489},
-//         {0.22261092,  0.31779705},
-//         {0.59410701,  0.1191493 },
-//         {0.1569244 ,  0.05827084},
-//         {0.61999389,  0.75780352},
-//         {0.10056238,  0.44661418},
-//         {0.9609345 ,  0.20627358},
-//         {0.86056549,  0.19085841}
-//     };
-//     std::cout << qhull2D(a);
-//     return 0;
-// }
