+++ components/mds/mds_cpp/qhull_2d.hpp
@@ -1,9 +1,7 @@
 #ifndef qhull_2d_hpp
 #define qhull_2d_hpp
 
-#ifndef armadillo
 #include <armadillo>
-#endif
 
 arma::mat qhull2D(arma::mat sample);
