+++ components/mds/mds_cpp/sammon.hpp
@@ -1,9 +1,7 @@
 #ifndef sammon_hpp
 #define sammon_hpp
 
-#ifndef armadillo
 #include <armadillo>
-#endif
 
 arma::mat euclid(arma::mat A, arma::mat B);
 std::pair<arma::mat, double> sammon(arma::mat D, int display = 2,
